plugin: add/rm dependencies due to resource usage
Problem: The priority plugin tracks resource usage across an
association's set of running jobs, but it does not enforce any limits
on this resource usage in the event where an association goes over
their limit.

Add a new method to the Association class to check to see if an
association's job would put them at or over their overall max
resources limit. Use this method in job.state.depend to determine
whether to hold the job. Use this method in job.state.inactive to
determine whether to release a held job from a per-association max
resources dependency.