add-user: add --default-project as an optional arg
Problem: The add-user command does not have an option to specify a
default project when adding an association for the first time. The
behavior currently is to pass a list of projects where the first one is
treated as the default. It would be convenient to have an option to also
just specify a default project by itself.

Add a new optional argument to the add-user command: --default-project,
which takes a string representing the name of the default project an
association will submit jobs under. Check for the existence of this
argument when determining the default project for an association before
just taking the first project listed in "projects" to keep backwards
compatibility.

If a default project is specified, make sure that it is also present in
the "projects" list, and if it is not, add it.