flux-account: add CLIMain decorator
Problem: The stack traces for errors raised in the flux-accounting
commands are not suppressed by default even though a lot of the commands
have custom error messages. This means that when an expected error
occurs in flux-accounting (like adding a user to a bank that does not
exist), the stack trace is very long and could be misleading to someone
running the command.

Add a decorator to main() from flux-core to enable suppression of large
stack traces that prepend a custom error message raised from
flux-accounting.