flux-account-service: condense Exception blocks
Problem: It was correctly pointed out in #597 that an exception caught
by the general "except Exception as exc" line could in fact be an
OSError, but a lot of these except blocks in flux-account-service still
have the "a non-OSError exception was caught" prefix.

Problem: There is a lot of duplicate code in flux-account-service.py in
the way that exceptions are handled, i.e. exceptions are handled one at
a time with the same prefix and same output of the specific exception
message. This code could be condensed into a general "except Exception"
block where the type of exception is just included in the exception
message.

Remove this "a non-OSError occurred" prefix from all functions in
flux-account-service.py that catch a general Exception.

Condense the various except blocks (except for KeyError except blocks,
which have to do with a missing key when actually calling the
appropriate command) into just one "except Exception as exc" block and
include the type of exception in the message.