service: add note about sqlite3 errors
Problem: There is a comment in the user bindings that explains that
sqlite3 errors and exceptions raised might not be the fault of the
programmer and includes a link to SQLite documentation. But since the
bindings are going to be updated to not have try/except blocks (since
the service handles exceptions now), this comment will be deleted.

Add the comment to the first exception block in the service in
view-user.