bank bindings: remove redundant try/except blocks
Problem: A lot of the bindings still have try/except blocks from when
flux-accounting did not have a systemd service and something that could
actually catch any exceptions raised from the functions themselves.
Thus, there is a lot of code in the bindings that looks like:

except Exception:
    raise Exception

Remove try/except blocks in the Python bindings that deal with bank
subcommands where appropriate since they are now handled in the
flux-accounting service.