job-usage: remove TRANSACTION for job usage update
Problem: The TRANSACTION that the job-usage update script employs
requests an exclusive lock on the database while it is running, which
means other reads/writes (even something as simple as "flux account
view-user") will hang until the transaction has finished running. An
admin should still be able to access the DB even when this script is
running on an interval.

Remove the declaration of a TRANSACTION when running the job-usage
update script and instead execute .commit() statements for every
association and bank update sequentially, which should allow other
reads/writes to be executed while this script is running.