list-banks: make table output default
Problem: The list-users command outputs in a table format by default,
which was discussed by multiple Flux developers to be the preferred
default (see #597). However, the list-banks command outputs in JSON by
default instead of in a table format.

Change the list-banks to output information in table format by default
and add --json as an optional argument.

Adjust the unit tests in t/python/t1008_banks_output.py to account for
the change in arguments to the list_banks() function.

Adjust the man page for list-banks to list --json as an optional
argument instead of --table.