list-queues: make table output default
Problem: The list-users command outputs in a table format by default,
which was discussed by multiple Flux developers to be the preferred
default (#597). However, the list-queues command outputs in JSON by
default instead of in a table format.

Change the list-queues to output information in table format by default
and add --json as an optional argument.

Adjust t1024-flux-account-queues.t to account for the change in
arguments to the list_queues() function.

Adjust the man page for list-queues to list --json as an optional
argument instead of --table.