update-usage: add INFO-level logging
Problem: The update-usage command (the script responsible for updating
job usage values) can take upwards of a minute or two to run because it
needs to update job usage values for every association and bank in the
database, i.e. executing an UPDATE statement in the database. This
requests a lock on the DB, and other flux-accounting commands executed
during this time might hang while this is running. But since there is
no logging of when this is running, figuring out why flux-accounting
commands are hanging can be confusing.

Add INFO level logging to the function called by the update-usage
command when the function starts/stops to give an idea of how long the
function takes to run.