update-usage: add missing update of current usage
Problem: In the case where no new jobs are found in a new half-life
period, the update-usage command does not update the "current" job
usage period for an association, which means that there is a discrepancy
between what the job_usage_factor_table shows for an association's job
usage and their total historical job usage.

Add an to update the current job usage period for an association in the
case where no new jobs are found in a new half-life period (i.e. their
current job usage should be 0).