t: add test for checking current usage
Problem: There is no test to see that the current usage for an
association is properly set in the case where there are no new jobs
found in a new half-life period.

Add a check in an existing test to ensure that the current job usage
for an association is 0.0 in the case where no new jobs are found in a
new half-life period.