plugin: set nnodes == 1 when only cores specified
Problem: The priority plugin has to estimate the resources requested by
a job in order to enforce max resources limits since the actual known
amount of resources used isn't known until the job gets to the
scheduler. So, when a job is submitted that only requests cores, the
helper function that returns the resource counts for the job will return
0 for the number of nodes, which is not technically the case - the job
will still use at least 1 node.

Update the increment_resources () and decrement_resources () functions
to set nnodes to 1 when ncores > 0 but nnodes == 0 in order to
accurately track resources used by a job. In job.state.depend, make the
same assumption when counting the resources for a prospective job.