t: update sharness test to expect nnodes == 1
Problem: There is a test in t1044-mf-priority-resource-limits.t that
counts the resources consumed by a job that only requests ncores. This
is no longer the case, however, as the plugin now sets nnodes == 1 for a
job that only requests cores.

Update this test to expect nnodes == 1 for a job that only specifies
ncores.