t: edit expected error message
Problem: There are a couple of tests in t1007-flux-account-users.t that
check the output of an error message when purposefully leaving out
required arguments. Now that these required arguments have a short
format, these expected error messages need to be updated to reflect the
change.

Update the expected error messages in t1007-flux-account-users.t to
account for the addition of short formats for the "username" and "bank"
arguments in the add-user command.