commands: add short options for queue commands
Problem: A number of commands that deal with queues in the
flux-accounting database do not offer short-form options, which results
in some inconvenient typing when specifying a lot of the optional args.
It would be nicer if these optional arguments had a short-form option.

Add some short options to the various commands that deal with queues.
Update the corresponding *-queue(1) pages with the new short options.