bank_table: add 'priority' column
Problem: The bank_table in the flux-accounting database does not have a
priority column, i.e. a value associated with each bank that could
affect a job's overall priority.

Add a new column to the bank_table: priority, a float associated with
each bank that could affect a job's overall priority if it is submitted
under that bank. Update the BANK_TABLE constant __init__.py and add
"priority" to the list of column headers. Update tests that look at the
column names in the bank_table to account for the addition of the
"priority" column.