accounting: update Association class
Problem: The Association class does not have a way to store the priority
associated with the bank they are running their job under, similar to
how the class already stores the priority associated with the queue they
are running their job under.

Update the Association class by adding a new member: "bank_factor", a
double that represents the priority associated with the bank they are
running their job under.

Update the unit tests for the Association class to account for the
addition of a new "bank_factor" member.