accounting: add Bank class
Problem: The priority plugin does not have a way to store certain
attributes on a per-bank basis, such as its associated priority.

Add a new class to the accounting header file called Bank, which will
store basic attributes about banks defined in the flux-accounting DB,
particularly their priority, if one is set.

Add a helper function to perform a lookup in a map of Bank objects and
retrieve a priority. If the bank cannot be found, just return a priority
of 0.0.