plugin: add unpack of bank information from DB
Problem: The priority plugin does not currently unpack bank attribute
information from the flux-accounting DB, but it will need to do this in
order to affect a job's priority calculation if it submitted under a
certain bank.

Add a section to the bulk update script to also send bank information,
particularly bank names and its associated priority to the plugin.

Unpack the bank information in the plugin and store it in a map that
contains Bank objects.

Add a "bank" initialization to the priority_weights map and an unpack
of an associated weight of the "bank" factor in the conf.update
callback.