plugin: unpack priority associated with banks
Problem: Banks can now have an associated floating-point priority
assigned to them, but they are not utilized anywhere in the plugin.

Unpack a priority associated with a bank and utilize it when calculating
the priority for a job. If the bank is updated while a job is still
pending, also unpack its associated priority to update the job's
priority.