doc: add note about active per-queue limits
Problem: There are no notes in flux-accounting documentation about
configuring max running jobs limits on a per-queue basis to be enforced
by the priority plugin.

Add a section in the flux-accounting guide making a note about how to
set a max running jobs limit for a queue and how it is enforced on an
association's set of running jobs.