service: use .get() instead of direct key access
Problem: The flux-accounting service uses direct dict key access to
fetch parameters for the various flux-accounting functions. This can be
problematic in the event that a key-value pair is not present in the
payload (such as an optional argument) and a KeyError will be raised.

Convert the direct key access in the payload dictionary for optional
arguments to instead use .get(), which will return None in the case that
a key-value pair does not exist in the dictionary.