cmd: move update-usage to its own script
Problem: The update-usage command is a part of the flux-accounting
service, which means that when it is running (especially on a large
flux-accounting DB that can take a lot of time), it will halt the
execution of other flux-accounting commands (even ones that just read
from the DB) until the update-usage command finishes running. This
command should be moved to its own script that establishes a separate
connection to the DB so other flux-accounting commands can run while
job usage values are being updated.

Create a new Python command script in the src/cmd/ directory called
flux-account-update-usage.py, a script that will just call
update_job_usage(), similar to what the update-usage command in the
flux-accounting service was doing.