t: update tests to reflect new update-usage script
Problem: The sharness tests in the flux-accounting testsuite still call
update-usage as a command that is part of the flux-accounting service,
but this is now its own script, i.e. the usage has changed to "flux
account-update-usage".

Update the call throughout the flux-accounting testsuite to reflect the
change to how job usage is updated, i.e. change
"flux account update-usage" to "flux account-update-usage".