flux-accounting guide: update cron script
Problem: The flux-accounting guide still refers to the update-usage
as part of the flux-accounting service, but this is now its own script,
i.e. the usage has changed to "flux account-update-usage".

Update the flux-accounting guide to reflect the change to how job usage
is updated, i.e. change "flux account update-usage" to
"flux account-update-usage".