db: add new priority_factor_weight table to DB
Problem: The multi-factor priority plugin defines default weights for
all of the different priority factors in the plugin code itself, making
it very inconvenient for other parts of flux-accounting (the DB and the
Python bindings in particular) to keep consistency with these defaults
if they are changed later on.

Begin to move this configuration of priority factor weights out of the
plugin by creating a new table in the flux-accounting DB called
priority_factor_weight_table, a table to store the integer weights of
the various factors used to calculate a job's priority in the priority
plugin. Update the existing tests that check for the different tables in
the DB to account for the addition of the new table. Update the DB
schema version since a new table is being added.