formatter: add PriorityFactorFormatter subclass
Problem: There is no subclass for printing out information about a
particular priority factor defined in the priority_factor_weight_table
as well as a custom error message, similar to the other formatter
subclasses.

Add a new PriorityFactorFormatter subclass, which inherits the
.as_json(), .as_table(), and .as_format_string() methods and uses its
own error message in the case that a specified factor does not exist in
the priority_factor_weight_table.