priorities: add list-factors command
Problem: There is no way to concisely list all of the defined priority
factors in the priority_factor_weights_table.

Add a new command called list-factors which will list all of the defined
priority factors in the priority_factor_weights_table. Add options to
view output in JSON format, with a format string, or customize the
columns displayed in output.