update-db: add insertion of priority factors in DB
Problem: The update-db command will update an old flux-accounting DB by
adding new tables and columns to tables, but does not insert any rows
into tables in the DB, but it will need to in the case where it is
updating a DB that never had the priority_factor_weight_table in it
beforehand because that table is created with values in the create-db
command; otherwise, the table will be created and it won't have any
data in it.

Add a function to flux-account-update-db.py to insert the priority
factors into the priority_factor_weight_table IF they do not exist
already. If they exist, ignore the insert since they might have a set
value other than the default and we do not want to override it.