doc: add page on limit configuration
Problem: There is a little description in the flux-accounting guide
about on limits in the priority plugin, but it is incomplete (it is
missing the max resources per-association limit) and could use some
more explanation and perhaps an example.

Add a new page to the flux-accounting docs about how limits work in
flux-accounting. Provide an example on the difference between soft
limits and hard limits. Reference the Limits page in the
flux-accounting guide.