plugin: send and unpack priority weight info
Problem: The flux-accounting database is now the central location for
storing the priority factor weights, but this information is not sent
to the plugin to actually be used when calculating the priority for a
job. This will be used *instead of* configuring the weights for each
factor in a TOML file and reloading config.

Remove the conf.update callback and replace it with a callback that
unpacks the priority factor weight information from the database.