t: change config of priority factor weights
Problem: The way priority factors weights are configured have changed
from being in a TOML file and reloaded with "flux config reload" to
being modified in the flux-accounting database and sent to the plugin
along with the other DB information, but there are tests in the
testsuite that still use the old method.

Edit the tests in the testsuite to use the new method of configuring
priority weights in the flux-accounting database and sending it with
"flux account-priority-update".