doc: remove flux-config-accounting(5)
Problem: The flux-accounting documentation still includes the man(5)
page for configuring priority weights in a TOML file and reloading
config, but this is no longer supported in favor of configuring the
priority factors in the database and sending it to the plugin along with
the rest of the flux-accounting DB information.

Remove the man(5) page from the documentation.