cmd: add new "jobs" command
Problem: There is no way to see a convenient breakdown of how various
job priorities for a user were calculated. Since there are multiple
factors with different weights, it might be helpful to list this
breakdown on the command line instead of making the user consult
documentation and calculate the priorities themselves.

Add a new command to flux-accounting: jobs, which takes one argument
(a username) and lists a breakdown of all of their jobs' priorities.
If no optional arguments are passed, all of the user's jobs will be
returned. However, this list can be filtered to show all jobs that have
run under a certain bank (by passing --bank) or show all jobs that have
run under a certain queue (by passing --queue).