priorities: add weight for "urgency" factor
Problem: When the flux accounting priority plugin is in use, the urgency
(described in RFC 30) has a very small impact on priority, which means
that users cannot meaningfully lower the priority of their jobs (other
than hold) relative to others, and sys admins cannot raise the priority
of jobs (other than expedite). This is because in the priority
calculation, the urgency is not multiplied by an integer weight like the
other factors, leaving it with little to no impact.

Add a weight to the "urgency" factor in the
priority_factor_weight_table and give it a default integer weight like
the other priority factors in the table.

Adjust one of the unit tests in t1011_priorities.py to account for the
addition of "urgency" in the priority_factor_weight_table.