plugin: unpack and store weight of "urgency"
Problem: The flux-accounting database defines a default weight and
now allows for the modification of a weight to be associated with the
"urgency" factor when calculating a job's priority, but it is not
actually used in the plugin.

Use the weight associated with the "urgency" factor when calculating a
job's priority by adding a product of the weight associated with the
"urgency" factor and the user (or admin) set urgency - 16.

Make the "urgency" factor nonrelevant in t1001-mf-priority-basic.t since
there are tests that submit jobs with custom urgencies and check
specific priority values.