doc: add table with weights for each factor
Problem: The flux-accounting guide makes mention of the default weights
for each priority factor, but as more factors are added (especially if
they have weights), it would be easier to consult a table that concisely
lists each priority factor and their associated default weight.

Add a table to the flux-accounting guide that lists each priority factor
and their default weight.