doc: add page on job priorities
Problem: There is no extensive documentation on how a job's priority is
calculated and how priorities can be affected by configuring different
factors, as well as viewing a breakdown of how a job's priority was
calculated with the "flux account jobs" command.

Add some documentation on job priorities by creating a new page that
goes into detail about priority calculation. Move some of the details
regarding priority calculation and configuration out of the
flux-accounting guide and into this new page. Reference this page in
the flux-accounting guide.