update-usage: remove extra .commit()
Problem: The apply_decay_factor() helper function has a .commit() call
at the end of the function, which will write all changes to disk.
However, this helper function is called in the middle of a SQLite
TRANSACTION, which will already call .commit() when the transaction is
finished running. Thus, this hinders performance of the update-usage
command.

Remove the extra .commit() from the end of apply_decay_factor().