view-job-records: add short option for --bank
Problem: There is no short option for the --bank optional argument in
view-job-records. In flux-core, -B is used to specify a bank.

Add -B as a short option for the --bank optional argument in the
view-job-records command.

Add a test in t1043-view-jobs-by-bank.t to filter jobs by a bank using
the short option.