service: use constant for checking DB schema
Problem: The flux-accounting service hard-codes the schema version of
the flux-accounting DB when checking to see if the DB is out-of-date and
needs to be updated. However, there is a constant that exists that will
remain up to date event when the schema version number is updated.

Use fluxacct.accounting.DB_SCHEMA_VERSION in place of a hard-coded
schema version number when checking whether the database needs to be
updated.