sql_util: add db_version() function
Problem: The account-priority-update and account-service files each have
their own check of the DB schema version of the flux-accounting
database. Thus, there is opportunity to create a function that can be
imported and used across the flux-accounting scripts. In the case of
flux-account-service.py, the check_db_version() function is not even
used.

Create a function which will just return the schema version number of
the flux-accounting DB, and use this function both in
flux-account-priority-update.py and flux-account-service.py.