jobs: use fair-share at priority calculation
Problem: The "jobs" command pulls the association's *current* fair-share
value instead of what their fair-share value was at the time of
job submission, which leaves inactive jobs with inaccurate priority
calculations in the output, particularly as an association's fair-share
value changes over time.

Pack the association's fair-share value at the time of priority
calculation using a memo event, and unpack this value later on when
creating rows of output for an association's jobs by looking at the
job's annotations.