Queue: add max_nodes_per_assoc as an attribute
Problem: The Queue class does not have an attribute for the max number
of nodes an association can have across all of their running jobs.

Add a max_nodes_per_assoc attribute. Give it the same default as the one
in the commands for adding and editing queues. Unpack it in the callback
for unpacking queue information from the flux-accounting database.