cmd: add edit-all-users command
Problem: There is no way to efficiently edit a column (or attribute) for
every association in association_table. This might come in handy in the
case that an admin wants to update something for every association, such
as changing the max_active_jobs limit from 7 to 100.

Add a new command to the flux-accounting command suite called
edit-all-users, which takes optional arguments for a number of columns
in association_table and executes an UPDATE statement for every
association in the table at once.