doc: update man(1) page for edit-user
Problem: The man(1) page for the edit-user command is missing a couple
of optional arguments. It also lists the --bank optional argument as a
modifiable attribute when it really should be listed as an option to
constrict an update to a specific row in association_table.

Update the man(1) page for the edit-user command to add the missing
optional arguments. Move the description of the --bank optional argument
to the top of the page to better clarify what its use is in the command.