plugin: enforce per-queue max nodes dependency
Problem: The priority plugin does not enforce a limit on the max number
of nodes an association can have across all of their running jobs in a
particular queue.

Add limit enforcement on the max number of nodes an association can have
across all of their running jobs in a particular queue.