view-user: add -J/--job-usage optional argument
Problem: An association's job usage value is made up of multiple job
usage values (i.e. it is the sum of both current and older job usage
factors that are stored in the job_usage_factor_table), but this data
is not viewable on the command line.

Add a new optional argument to the view-user command: -J/--job-usage,
which will return all the data stored for an association in
job_usage_factor_table. Make it mutually exclusive with the --fields
optional argument.