doc: move images/ folder up one level
Problem: There is an images/ folder that lives *just* in the guide/
folder that contains the icon for the LGPL-3.0 icon. As more and more
images get added to the flux-accounting documentation, it would be nice
if all of these images could be stored in a more accessible location
for any section or subsection of the flux-accounting documentation to
refer to.

Move the images/ folder up one level.