doc: add module structure document
Problem: flux-accounting, as a whole, is composed of multiple modules
that allow it to run alongside a Flux instance - a SQLite database,
multiple scripts that run in the background, and a systemctl service
that interfaces with this database. It would be helpful to actually
diagram these different modules and how they interface with each other.

Add a "Module Structure" page to the flux-accounting documentation with
a number of diagrams explaining how the different components of
flux-accounting work together to run flux-accounting as a service.