job: set nnodes = 1 when ncores > 0, nnodes == 0
Problem: The jj_get_counts_json () function will leave nnodes == 0 if
a user specifies cores but not nodes, even though this is not accurate;
a specification of cores will use *at least* one node. The priority
plugin sets this default outside of the count_resources () method for
the Job class, but it would make more sense to put this within the
method itself.

Move this assignment of nnodes = 1 to inside the count_resources ()
method.