plugin: add function for joining string of vectors
Problem: In order to provide more context for an error message as to
why an association can't run a job under a certain queue or project,
the list of the association's valid projects or queues should also be
returned. Because this list is stored as a vector in the Association
class, there needs to be a way to join these strings into just one
comma-delimited string.

Add a helper function to the plugin for combining a string of vectors
into just one string that takes a custom delimiter.