plugin: drop iterator when performing map lookup
Problem: The map lookup of a queue in the queues map in the plugin
defines an iterator to store the result of the lookup, but this
lookup can just be directly compared to the .end () of the map.

Drop the assignment of the map lookup to an iterator in favor of just
directly comparing the result of the lookup within the if-statement.