database: add duration columns to "jobs" table
Problem: The requested duration for jobs submitted by associations is
stored in jobspec, but it is not conveniently stored anywhere for use
throughout flux-accounting. It would be nice if this value was easily
available for tracking on a per-association or other basis. Storing the
actual duration of the job would also be useful for tracking purposes.

Add new columns to the jobs table to track the requested and actual
durations of jobs. Update the schema version of the flux-accounting DB
to account for the addition of the new column to the "jobs" table.