cmd: remove --output-file argument
Problem: The flux-accounting service offers an --output-file argument to
direct output of any of the commands to a file, but output can already
be redirected to a file with the ">" character.

Remove the --output-file argument from the flux-accounting command
suite.

Remove the write_records_to_file () function from the
jobs_table_subcommands file.

Remove the one test in t1011-job-archive-interface.t that checks
redirecting output to a file with the --output-file argument.