view-job-records: add --requested-duration option
Problem: There is no way to filter the jobs table by requested_duration.

Add a new optional argument to the view-job-records command that can
take one or more filters for the requested_duration column and filter
the jobs table. Add a helper function to jobs_table_subcommands.py to
validate the expressions passed in to make sure they are valid filters
for the column before adding them to the SELECT query.