t: add helper script for updating actual_duration
Problem: There is no convenient way to spoof the actual duration of jobs
in the testsuite so that you don't have to actually let the job run for
the desired duration.

Add a helper script that just updates the actual duration of a job in
the jobs table to be used in the testsuite.