open_db (): open the DB in robust mode
Problem: The open_db () function in data_writer_db.cpp opens the
flux-accounting DB in with SQLITE_OPEN_READWRITE but does not set any
busy policy or Write-Ahead Logging (WAL) mode, so even a brief overlap
between a writer and a reader will return SQLITE_BUSY, and the fairshare
update will not go through.

Add a busy timeout to open_db () to wait for the DB in the event it is
locked at the moment. Set PRAGMAs for enabling write-ahead logging and
synchronous modes when writing to the database. Set a PRAGMA for storing
temporary tables and indices in memory rather than on desk when updating
the fair-share values.