bind_param (): make safe for temporaries
Problem: In sqlite3_bind_text (), NULL is passed as the destructor,
which tells SQLite that the pointer must remain valid until it is
finalized, but these pointers are bound from temporaries in
update_fairshare_values (), which can lead to errors.

Pass SQLITE_TRANSIENT for the destructor, which tells SQLite to make
its own copy of the value before it returns from the function call so
that SQLite has a stable value for a value that might change, such as
the fair-share value.