write_acct_info (): use explicit WRITE transaction
Problem: Currently, each row UPDATE effectively is its own transaction
on the database, which can increase the block it has on other operations
throughout flux-accounting that might interact with the database.

Wrap the UPDATE statements to the fair-share values for each association
in a single transaction to reduce the potential of a lock. Institute a
ROLLBACK in the event the transaction fails to restore the database to
its state before the transaction began.