scrub-old-jobs: add return code
Problem: The scrub_old_jobs() function does not return anything at
function end, so the flux-accounting service will return and output
"None" in response, which is confusing for a successful call of the
function.

Add a 0 return code to the end of scrub_old_jobs() when the function
runs successfully.