update_job_usage(): use row names
Problem: The update_job_usage() function uses tuple indices, which makes
it not immediately clear what is being passed to the calc_usage_factor()
function. The sqlite3 module offers a sqlite3.Row property to be set on
the SQLite Connection object to be able to use row names instead of just
the indices of the tuple.

Use the sqlite3.Row property on the Connection object in order to be
able to use the row names of the SQLite query instead of tuple indices.