update-usage: condense SELECT queries
Problem: calc_usage_factor() issues a SELECT query in the case where it
finds new jobs in the same half-life period for an association from a
table it is already querying from before the function is called. There
is opportunity to remove this SELECT query in favor of storing this
value *ahead* of calling calc_usage_factor() and passing it as an
argument to the function, where it can be used if needed.

Remove the get_curr_usg_bin() function in favor of just fetching
usage_factor_period_0 in the initial query ahead of calling
calc_usage_factor() for every association, and pass it as a function
argument. Adjust the unit tests in python/t1006_job_archive.py to
account for the new function argument passed to calc_usage_factor().